+++ Examples/004-Structs-In-Packets/StructsInPackets.cs
@@ -8,7 +8,7 @@ namespace Link.Examples._004_Structs_In_Packets;
 public static class StructsInPackets
 {
     private const string IpAddress = "127.0.0.1";
-    private const int Port = 7777;
+    private const ushort Port = 7777;
 
     public static void Main()
     {
