+++ Examples/010-Connection-Statistics/ConnectionStatistics.cs
@@ -11,7 +11,7 @@ namespace Link.Examples._010_Connection_Statistics;
 public static class ConnectionStatistics
 {
     private const string IpAddress = "127.0.0.1";
-    private const int Port = 7777;
+    private const ushort Port = 7777;
 
     public static void Main()
     {
